ref: split transform functions
